tools/adr/requirements.txt
jsonschema>=4.0.0
pyyaml>=6.0

